Upgrade deps, restyle toolbar
