+++ package.json
@@ -37,48 +37,43 @@
     "*": "pnpm check:fix"
   },
   "dependencies": {
-    "@cantoo/pdf-lib": "^2.4.1",
+    "@cantoo/pdf-lib": "^2.4.2",
     "@convex-dev/better-auth": "0.7.0-alpha.1",
     "@convex-dev/migrations": "^0.2.9",
     "@faker-js/faker": "^9.9.0",
     "@maskito/core": "^3.10.0",
     "@maskito/react": "^3.10.0",
-    "@namesake/tiptap-extensions": "^3.0.0",
+    "@namesake/tiptap-extensions": "^3.0.1",
     "@radix-ui/colors": "^3.0.0",
     "@react-email/components": "0.1.1",
     "@tailwindcss/vite": "^4.1.11",
-    "@tanstack/react-router": "^1.125.4",
-    "@tiptap/core": "^2.25.0",
-    "@tiptap/extension-blockquote": "^2.25.0",
-    "@tiptap/extension-bold": "^2.25.0",
-    "@tiptap/extension-bullet-list": "^2.25.0",
-    "@tiptap/extension-document": "^2.25.0",
-    "@tiptap/extension-gapcursor": "^2.25.0",
-    "@tiptap/extension-hard-break": "^2.25.0",
-    "@tiptap/extension-heading": "^2.25.0",
-    "@tiptap/extension-history": "^2.25.0",
-    "@tiptap/extension-italic": "^2.25.0",
-    "@tiptap/extension-link": "^2.25.0",
-    "@tiptap/extension-list-item": "^2.25.0",
-    "@tiptap/extension-list-keymap": "^2.25.0",
-    "@tiptap/extension-ordered-list": "^2.25.0",
-    "@tiptap/extension-paragraph": "^2.25.0",
-    "@tiptap/extension-placeholder": "^2.25.0",
-    "@tiptap/extension-text": "^2.25.0",
-    "@tiptap/extension-typography": "^2.25.0",
-    "@tiptap/pm": "^2.25.0",
-    "@tiptap/react": "^2.25.0",
+    "@tanstack/react-router": "^1.127.3",
+    "@tiptap/core": "^3.0.1",
+    "@tiptap/extension-blockquote": "^3.0.1",
+    "@tiptap/extension-bold": "^3.0.1",
+    "@tiptap/extension-document": "^3.0.1",
+    "@tiptap/extension-hard-break": "^3.0.1",
+    "@tiptap/extension-heading": "^3.0.1",
+    "@tiptap/extension-italic": "^3.0.1",
+    "@tiptap/extension-link": "^3.0.1",
+    "@tiptap/extension-list": "^3.0.1",
+    "@tiptap/extension-paragraph": "^3.0.1",
+    "@tiptap/extension-text": "^3.0.1",
+    "@tiptap/extension-typography": "^3.0.1",
+    "@tiptap/extensions": "^3.0.1",
+    "@tiptap/pm": "^3.0.1",
+    "@tiptap/react": "^3.0.1",
     "@zxcvbn-ts/core": "^3.0.4",
     "@zxcvbn-ts/language-common": "^3.0.4",
     "@zxcvbn-ts/language-en": "^3.0.2",
     "better-auth": "1.2.12",
     "convex": "^1.25.2",
-    "convex-helpers": "^0.1.96",
+    "convex-helpers": "^0.1.99",
     "language-name-map": "^0.3.0",
     "lucide-react": "^0.525.0",
-    "motion": "^12.23.0",
+    "motion": "^12.23.3",
     "postcss": "^8.5.6",
-    "posthog-js": "^1.256.2",
+    "posthog-js": "^1.257.0",
     "react": "^19.1.0",
     "react-aria": "^3.41.1",
     "react-aria-components": "^1.10.1",
@@ -88,32 +83,32 @@
     "react-random-reveal": "^2.0.1",
     "resend": "4.6.0",
     "sonner": "^2.0.6",
-    "storybook": "^9.0.15",
+    "storybook": "^9.0.16",
     "tailwind-variants": "^1.0.0",
-    "tailwindcss": "^4.0.14",
+    "tailwindcss": "^4.1.11",
     "tw-animate-css": "^1.3.5",
     "typed-usa-states": "^2.1.0",
     "use-debounce": "^10.0.5",
-    "zod": "^3.25.74"
+    "zod": "^3.25.76"
   },
   "devDependencies": {
     "@axe-core/playwright": "^4.10.2",
     "@babel/preset-react": "^7.27.1",
-    "@biomejs/biome": "^2.0.6",
+    "@biomejs/biome": "^2.1.1",
     "@changesets/cli": "^2.29.5",
     "@edge-runtime/vm": "^5.0.0",
     "@emotion/is-prop-valid": "^1.3.1",
-    "@playwright/test": "^1.53.2",
+    "@playwright/test": "^1.54.1",
     "@react-aria/test-utils": "1.0.0-alpha.8",
-    "@storybook/addon-themes": "^9.0.15",
-    "@storybook/react-vite": "^9.0.15",
-    "@tanstack/router-devtools": "^1.125.4",
-    "@tanstack/router-plugin": "^1.125.5",
+    "@storybook/addon-themes": "^9.0.16",
+    "@storybook/react-vite": "^9.0.16",
+    "@tanstack/router-devtools": "^1.127.3",
+    "@tanstack/router-plugin": "^1.127.5",
     "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.3.0",
     "@testing-library/user-event": "^14.6.1",
-    "@types/node": "^24.0.10",
+    "@types/node": "^24.0.13",
     "@types/react": "^19.1.8",
     "@types/react-dom": "^19.1.6",
     "@vitejs/plugin-react-swc": "^3.10.2",
@@ -126,11 +121,11 @@
     "husky": "^9.1.7",
     "jsdom": "^26.1.0",
     "lint-staged": "^16.1.2",
-    "playwright": "^1.53.2",
+    "playwright": "^1.54.1",
     "prop-types": "^15.8.1",
     "react-email": "4.0.17",
     "rollup-plugin-webpack-stats": "^2.1.0",
-    "sharp": "^0.34.2",
+    "sharp": "^0.34.3",
     "svgo": "^4.0.0",
     "tailwind-merge": "^3.3.1",
     "tailwindcss-react-aria-components": "^2.0.0",
