+++ src/components/common/Toolbar/Toolbar.tsx
@@ -6,7 +6,7 @@ import {
 import { tv } from "tailwind-variants";
 
 const styles = tv({
-  base: "flex gap-1",
+  base: "flex gap-0.5",
   variants: {
     orientation: {
       horizontal: "w-full flex-row items-center justify-start overflow-x-auto",
